added remaining requestion location test files
