+++ body_test.go
@@ -12,15 +12,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func setupRouter() *gin.Engine {
-	gin.SetMode(gin.TestMode)
-	r := gin.Default()
-	r.GET("/ping", func(c *gin.Context) {
-		c.String(200, "pong")
-	})
-	return r
-}
-
 func TestBodyValidationChain(t *testing.T) {
 	body := `{
 		"name": {"first": "Tom", "last": "Anderson"},
